+++ Source/uc_asset_manager.py
@@ -27,8 +27,13 @@ def logout():
         ucam.identity.user_login.logout(True)
 
 
-def export_file_with_preview(path: str, preview_path: str, name: str, description: str,
-                             tags_list: List[str], org_id: str, project_id: str):
+def __upload_file_to_dataset(org_id: str, project_id: str, asset_id: str, version: str, dataset_id: str, path: str):
+    upload_asset = AssetFileUploadInformation(org_id, project_id, asset_id, version, dataset_id, path)
+    ucam.assets.upload_file(upload_asset)
+
+
+def create_asset(path: str, preview_path: str, name: str, description: str,
+                 tags_list: List[str], org_id: str, project_id: str) -> str:
     asset_creation = AssetCreation(name,
                                    description,
                                    ucam.assets.asset_type.MODEL_3D,
@@ -37,12 +42,32 @@ def export_file_with_preview(path: str, preview_path: str, name: str, descriptio
     asset_id = ucam.assets.create_asset(asset_creation, org_id, project_id)
     version = '1'
     datasets = ucam.assets.get_dataset_list(org_id, project_id, asset_id, version)
-    upload_asset = AssetFileUploadInformation(org_id, project_id, asset_id, version, datasets[0].id, path)
-    ucam.assets.upload_file(upload_asset)
+    __upload_file_to_dataset(org_id, project_id, asset_id, version, datasets[0].id, path)
     if preview_path is not None:
-        upload_preview_file = AssetFileUploadInformation(org_id, project_id, asset_id,
-                                                         version, datasets[1].id, preview_path)
-        ucam.assets.upload_file(upload_preview_file)
+        __upload_file_to_dataset(org_id, project_id, asset_id, version, datasets[1].id, preview_path)
+    ucam.interop.open_browser_to_asset_details(org_id, project_id, asset_id, version)
+    return asset_id
+
+
+def update_asset(path: str, preview_path: str, name: str, description: str,
+                 tags_list: List[str], org_id: str, project_id: str, asset_id: str):
+    version = '1'
+    asset_update = AssetUpdate(name,
+                               description,
+                               ucam.assets.asset_type.MODEL_3D,
+                               tags_list)
+    if ucam.assets.update_asset(asset_update, org_id, project_id, asset_id, version):
+        datasets = ucam.assets.get_dataset_list(org_id, project_id, asset_id, version)
+
+        for asset_dataset in datasets:
+            file_list = ucam.assets.get_file_list(org_id, project_id, asset_id, version, asset_dataset.id)
+            for file in file_list:
+                ucam.assets.remove_file(org_id, project_id, asset_id, version, asset_dataset.id, file.path)
+
+        __upload_file_to_dataset(org_id, project_id, asset_id, version, datasets[0].id, path)
+        if preview_path is not None:
+            __upload_file_to_dataset(org_id, project_id, asset_id, version, datasets[1].id, preview_path)
+
     ucam.interop.open_browser_to_asset_details(org_id, project_id, asset_id, version)
 
 
@@ -52,6 +77,10 @@ def get_organizations() -> List[Organization]:
 
 def get_projects(org_id: str) -> List[Project]:
     return ucam.identity.get_project_list(org_id)
+
+
+def get_assets(org_id: str, project_id: str) -> List[Asset]:
+    return ucam.assets.get_asset_list(org_id, project_id)
 
 
 def uninitialize():