+++ Source/uc_blender_utils.py
@@ -17,9 +17,9 @@ def _get_preview_image(file_path):
     bpy.ops.render.render(write_still=True)
 
 
-def uc_addon_execute(org_id: str, project_id: str, name: str, description: str,
-                     tags_list: List[str], generate_preview: bool):
-    from .uc_asset_manager import export_file_with_preview
+def uc_create_asset(org_id: str, project_id: str, name: str, description: str,
+                    tags_list: List[str], generate_preview: bool) -> str:
+
     temp_dir = tempfile.mkdtemp()
 
     temp_fbx_file = os.path.join(temp_dir, f"{name}.fbx")
@@ -31,6 +31,25 @@ def uc_addon_execute(org_id: str, project_id: str, name: str, description: str,
             preview_file = os.path.join(temp_dir, "thumbnail.png")
             _get_preview_image(preview_file)
 
-        export_file_with_preview(temp_fbx_file, preview_file, name, description, tags_list, org_id, project_id)
+        from .uc_asset_manager import create_asset
+        return create_asset(temp_fbx_file, preview_file, name, description, tags_list, org_id, project_id)
+    finally:
+        shutil.rmtree(temp_dir)
+
+
+def uc_update_asset(org_id: str, project_id: str, asset_id: str, name: str, description: str, tags_list: List[str],
+                    generate_preview: bool):
+    temp_dir = tempfile.mkdtemp()
+    temp_fbx_file = os.path.join(temp_dir, f"{name}.fbx")
+    try:
+        _export_fbx(temp_fbx_file)
+
+        preview_file = None
+        if generate_preview:
+            preview_file = os.path.join(temp_dir, "thumbnail.png")
+            _get_preview_image(preview_file)
+
+        from . import uc_asset_manager
+        uc_asset_manager.update_asset(temp_fbx_file, preview_file, name, description, tags_list, org_id, project_id, asset_id)
     finally:
         shutil.rmtree(temp_dir)